refactor: 파일 index 추가 및 정리
