+++ frontend/src/hooks/review/index.ts
@@ -1,5 +1,9 @@
-export { default as useReviewFavorite } from './useReviewFavorite';
 export { default as useDisplayTag } from './useDisplayTag';
+export { default as useFormData } from './useFormData';
+export { default as useReviewFavorite } from './useReviewFavorite';
+export { default as useReviewImageUploader } from './useReviewImageUploader';
+export { default as useReviewRegisterForm } from './useReviewRegisterForm';
 export { default as useReviewTags } from './useReviewTags';
-export { default as useSelectedTags } from './useSelectedTags';
 export { default as useReviewTextarea } from './useReviewTextarea';
+export { default as useSelectedTags } from './useSelectedTags';
+export { default as useStarRate } from './useStarRate';
